Add initial Swagger API definition
This adds an initial Swagger API definition YAML file. In its current
state this is to be treated as a mere proposal. It defines a few needed
object types as well as a few API endpoints.
This is nowhere near final. Additions and modifications are more than
welcome, especially when it comes to how the parameters are being passed
to the endpoints, the possible responses, and/or any other missing
functionality.