Add post method to the “statuses” route
This adds the “post” method to the statuses route allowing the user to
check in. It also adds these tags:
* “Statuses” groups methods of the statuses route
* “Check-in” groups all methods required for a complete check-in workflow

Additionally, the “trip” property of the Status object definition is
being renamed to “journey” to be consistent with other definitions. Also,
the maximum length of the “statusText” property is defined to be 280.