Add caching to user, follows, leaderboard routes
This adds the optional If-Modified-Since header to the get method for
the mentioned routes. This allows for client-side caching and traffic
reduction. It gives the server the opportunity to respond with a 304
status code (== not modified) and an empty body.
Accordingly, Last-Modified headers are added to the 200 responses for
all of these routes.