Add routes to modify and delete a status
Adds the /status/{id} route with “put” and “delete” methods. This allows
a user to modify the status text of one particular status, or to delete
a status altogether.
These operations are defined to work on one status only as they are
simple enough to be repeated by the user later in case they should fail
due e.g. to bad cell reception.