Add retrieval of a user’s statuses
Since we want to display the statuses of a user on their profile page
we need a way to retrieve them. This adds the value “user” to the “view”
parameter’s enum in the /statuses route. It also adds a parameter which
will take a userId and must be present if the “view” parameter was set
to “user”.